Create mart.attack.vision naming space for vision-related components.
