+++ mart/attack/__init__.py
@@ -8,3 +8,4 @@ from .initializer import *
 from .objective import *
 from .perturber import *
 from .projector import *
+from .vision import *
