--- mart/attack/composer/__init__.py
@@ -1,2 +1,0 @@
-from .modular import *
-from .patch import *
