--- mart/attack/objective/__init__.py
@@ -1,3 +1,0 @@
-from .base import *
-from .classification import *
-from .object_detection import *
