+++ mart/attack/vision/__init__.py
@@ -1,0 +1,6 @@
+from ...utils.imports import _HAS_TORCHVISION
+
+if _HAS_TORCHVISION:
+    from .composer import *
+    from .initializer import *
+    from .objective import *
